+++ renku/domain_model/workflow/provider.py
@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright 2017-2022 - Swiss Data Science Center (SDSC)
+# Copyright 2017-2023 - Swiss Data Science Center (SDSC)
 # A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
@@ -40,7 +39,7 @@ class IWorkflowProvider(metaclass=ABCMeta):
         pass
 
     @abstractmethod
-    def workflow_execute(self, dag: "nx.DiGraph", basedir: Path, config: Dict[str, Any]):
+    def workflow_execute(self, dag: nx.DiGraph, basedir: Path, config: Dict[str, Any]):
         """Executes a given ``AbstractPlan`` using the provider.
 
         Returns: