+++ renku/ui/service/controllers/workflow_plans_list.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Copyright 2020 - Swiss Data Science Center (SDSC)
 # A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
@@ -34,7 +33,7 @@ class WorkflowPlansListCtrl(ServiceCtrl, RenkuOperationMixin):
     def __init__(self, cache, user_data, request_data):
         """Construct a plans list controller."""
         self.ctx = WorkflowPlansListCtrl.REQUEST_SERIALIZER.load(request_data)
-        super(WorkflowPlansListCtrl, self).__init__(cache, user_data, request_data)
+        super().__init__(cache, user_data, request_data)
 
     @property
     def context(self):