+++ CHANGES.rst
@@ -1,5 +1,5 @@
 ..
-    Copyright 2017-2022 - Swiss Data Science Center (SDSC)
+    Copyright 2017-2023 - Swiss Data Science Center (SDSC)
     A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
     Eidgenössische Technische Hochschule Zürich (ETHZ).
 
@@ -17,6 +17,55 @@
 
 Changes
 =======
+
+`2.3.2 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.3.1...v2.3.2>`__ (2023-03-13)
+-------------------------------------------------------------------------------------------------------
+
+Bug Fixes
+~~~~~~~~~
+
+-  **core:** Fix a bug that caused crashing when migrating to v10 metadata
+   (`f9da0be <https://github.com/SwissDataScienceCenter/renku-python/commit/f9da0be8585a692c49fbdaa19316b779a965db03>`__)
+
+`2.3.1 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.3.0...v2.3.1>`__ (2023-03-10)
+-------------------------------------------------------------------------------------------------------
+
+Bug Fixes
+~~~~~~~~~
+
+-  **cli:** set git credentials helper correctly on git clone
+   (`#3348 <https://github.com/SwissDataScienceCenter/renku-python/issues/3348>`__)
+   (`ac21c32 <https://github.com/SwissDataScienceCenter/renku-python/commit/ac21c3266246f011ec0bfd785e53fd03a553eee4>`__)
+-  **core:** update template checksum when updating Dockerfile
+   (`#3351 <https://github.com/SwissDataScienceCenter/renku-python/issues/3351>`__)
+   (`24407e5 <https://github.com/SwissDataScienceCenter/renku-python/commit/24407e5857faab0fe4d20883039f2a831977b15c>`__)
+-  **core:** parallelize dataset add to improve performance
+   (`#3338 <https://github.com/SwissDataScienceCenter/renku-python/issues/3338>`__)
+   (`85ab141 <https://github.com/SwissDataScienceCenter/renku-python/commit/85ab14199c83815f6d3a63fd6dc721638b3b4c95>`__)
+
+`2.3.0 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.2.0...v2.3.0>`__ (2023-02-27)
+-------------------------------------------------------------------------------------------------------
+
+Bug Fixes
+~~~~~~~~~
+
+-  **cli:** fix renku workflow visualize crashing for large graphs due
+   to a curses error
+   (`#3273 <https://github.com/SwissDataScienceCenter/renku-python/issues/3273>`__)
+   (`5898cbb <https://github.com/SwissDataScienceCenter/renku-python/commit/5898cbb5756c86013c169e234e8002285e4e84a7>`__)
+-  **core:** metadata v10 migration fixes
+   (`#3326 <https://github.com/SwissDataScienceCenter/renku-python/issues/3326>`__)
+   (`38aa53c <https://github.com/SwissDataScienceCenter/renku-python/commit/38aa53c3fa2a15976ff1cce68b5a21ca24df2078>`__)
+-  **svc:** properly use Redis sentinel
+   (`#3319 <https://github.com/SwissDataScienceCenter/renku-python/issues/3319>`__)
+   (`4a52069 <https://github.com/SwissDataScienceCenter/renku-python/commit/4a5206971fb8212a399d06132fe811d948edb96f>`__)
+
+Features
+~~~~~~~~
+
+-  **core:** add support for ssh connections into sessions
+   (`#3318 <https://github.com/SwissDataScienceCenter/renku-python/issues/3318>`__)
+   (`c024650 <https://github.com/SwissDataScienceCenter/renku-python/commit/c024650c96beef6d1ae2153b18b5b142171f2d80>`__)
 
 `2.2.0 <https://github.com/SwissDataScienceCenter/renku-python/compare/v2.1.0...v2.2.0>`__ (2023-02-10)
 -------------------------------------------------------------------------------------------------------